refactor: split artifact taxonomy and observation roots
Separate schema observation models, identity helpers, and runtime extraction from the public root, and do the same for raw artifact taxonomy classification.
